Adds basic interface for recordings
Currently display a basic interface for recordings, just to be used in
dev, this needs to be polished but it'll serve as the basis for the
recordings interface.

It includes support for dialect in recordings

Now recordings support dialects so we can have more than one recording
for each term with a different dialect, the dialect is an association
with Subjects.

Added some fixes to the CSS to improve the general markup.

The tree link was missing the %%ID%% identifier so when it was clicked
it didn't redirect properly.

Update engine version for merge into master