Added non-definition data models and specs
Added a dummy app that uses Rails5.2 to do the testing, it is currently
using Rails 5.2 because Terms depends on ActiveStorage, so when we add
the tests to that part of the engine we support it on the dummy app.

Added the basic model for Model Sentences, Currently because a Term and
a definition could potentially have a model sentence we decided to use a
polymorphic association to represent it, and build a module that can be
included to make a Term or a definition handle Model sentences, the new
Module is:

HasModelSentence

It creates the relationship between the model (called Context) and the
Model Sentence.

Made setting up initial database easier. Now simply running:

rake app:dictionary_to_terms:db:prepare RAILS_ENV=test

will make the environment ready for running "rspec". Make sure to include database.yml and flare.yml.

Added binstub for rake and fixed model sentence spec

Added the bin/stub for rake so we have everything running from teh
correct Rake.

Now we can run rake and rails from the binstubs and have the environment
be consistent.

Add Passage model

Refactor tests to get value from seed data instead of creating new language every run

Cleaned up code and added new models

+ Removed unused test directory.

  We are now using RSpec so no need to keep the old empty test directory

+ Added the pull request template.

  This to help improve our pull requests comments on github.

+ Added Model Sentence

  We can now have model sentences associated to definitions (via the
  HasModelSentences module), it is flexible enough to have the model
  sentece associated directly to term but its not currently being
  associated we would have to change the feature model to do that.

+ Added Passages

  Now we can associate Passages to a definition or a term, currently the
  definition is the only one that includes the module(HasPassages) that
  adds the passage but it could be added directly to features if we
  want.

+ Added Etymology support

  Now terms have support for Etymology, Etymology can have types of
  etymology that are represented by a Etymology subject association, this
  means that we create for an etymolgy a etymology_subject_association
  that has branch_id: 182 (this is the subject id for "Etymology type")
  and then we can add to the subject_id: any of the current subjects for
  etymology types (Basic syllabic[202], Creative[204], Historical[203])

An Etymology will have a type and this type will be represented by an
association to subjects, the decision to differentiate this subject
association from the regular is that we are restricting it to be only
one and the branch_id to the specific association is set by default.

Add annotate to models and specs