Updated the engine for new end-user interface
Built the new end-user interface for term, this includes:
+ Recordings
+ Editorial interface has now the style of the whole site
+ Change color schema of the whole application

Additional/details on changes included in this commit:

+ Fixed asset initializers and admin layout
+ Fixed the assets initializer to include the vendor directory and use the
shanti-main-terms.css.
+ Corrected the admin layout to set the application page to page-terms, used
on the CSS to target only code for terms.
+ Now there is only one partial for the whole recordings, they will be
displayed as a list and you can select one to be played. Added a button
to do the play but we need to style it.
+ Change recording display to not reload every time we play, only when
changing the selector.
+ Remove autocollapse option to fly-out tree
+ Updated engine version for merge into master
+ Support changes on kmaps_engine helper method feature_name_header (https://github.com/shanti-uva/kmaps_engine/pull/29)