Added JSON and XML API endpoints for Recordings
Add the new API endpoint so Mandala can obtain a list of recordings for
any given term. The API can return XML or JSON depending on the
extension in the request.

Examples:
http://localhost/terms/features/11/recordings.xml
http://localhost/terms/features/11/recordings.json
http://localhost/terms/features/11/recordings/6.json
http://localhost/terms/features/11/recordings/6.xml
http://localhost/terms/recordings/6.json
http://localhost/terms/recordings/6.xml

We are using jbuilder in order to build the JSON object and builder
for XML.

Update engine version for merge into master