Add Definition support for multiple authors
I created an intermediate table `authors_definitions` to keep the
relation between definitions and authors. The migration has to be run on
Terms, and update the Schema. Same for tests.

Also had to overwrite the definitions_controller `update` because
ActiveResource wouldn't be able to handle the change in `author_ids`
properly.

Based my solution on kmaps_engine descriptions authors associations.