Editorial interface for definition assiciation/manu 6535/drc2r (#47)
Adds new definition association interface to term relationship

* Adds support for feature relations to create in a single form instead of just term to term relation, term to definition or definition to definition relationship.
* Adds definition associations controller with implemented edit, show and delete.
* Displays associated definitions in the end user interface.

When doing a new feature relation, if a term to term relationship is selected a feature_relation is created. If a term to definition, definition to definition or definition to term relationship is selected a definition association is created. This definition association can be edited or deleted under the definition show. Editing a definition association allows the editor to switch from the leading definition to another definition of the same term, as changing the association to the associated term or another from that associated term.

* Adds definition association model, migration and tests
* Adds definition capabilities to feature relation form fields
* Adds definition association link to definitions
* Adds new Definition association interface to term relationship
* Adds definition association model, migration and tests
* Adds definition capabilities to feature relation form fields
* cleaning up whitespace